retour sur les modification ajout de commentaires et correction

+++ Bam2peut-etre.py
@@ -23,7 +23,7 @@ def creer_base(h_ouverture : int, min_ouverture : int, h_fermeture : int, min_fe
         min_ouverture INT,
         h_fermeture INT,
         min_fermeture INT,
-        stock_1place INT CHECK (stock_1place <= 50), -----Max 50 kayaks de chaque type
+        stock_1place INT CHECK (stock_1place <= 50), -----Max 50 kayaks de chaque type -----
         stock_2place INT CHECK (stock_2place <= 50)
         )
     """)
@@ -31,7 +31,9 @@ def creer_base(h_ouverture : int, min_ouverture : int, h_fermeture : int, min_fe
     cur.execute("INSERT INTO boutique_location VALUES (?, ?, ?, ?, ?, ?)", (h_ouverture, min_ouverture, h_fermeture, min_fermeture, nb_1place, nb_2places))
 
     # --- Création de la table calendrier ---
-    #On initialise en 01/01/2026 ? C'est dangereux d'appeler une table 'date' sachant que date est un type en SQLite.
+    #On initialise en 01/01/2026 ?
+    # g fait ca au pif 
+    # C'est dangereux d'appeler une table 'date' sachant que date est un type en SQLite. ------------- Vrai j'y avait pas réflechi
     cur.execute("""
     CREATE TABLE IF NOT EXISTS calendrier( 
         annee INT,
@@ -41,7 +43,7 @@ def creer_base(h_ouverture : int, min_ouverture : int, h_fermeture : int, min_fe
     """)
     cur.execute("INSERT INTO calendrier VALUES (2026, 1, 1)")
 
-    # --- Création de la table client ---
+    # --- Création de la table client --- ------------------ je ne sais pas si on doit s'occuper de la partie gestion des clients mais je suis d'accord que c'est important
     cur.execute("""
     CREATE TABLE IF NOT EXISTS client(
         id_client INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -52,11 +54,14 @@ def creer_base(h_ouverture : int, min_ouverture : int, h_fermeture : int, min_fe
 
     # --- Création de la table kayak ---
     #On identifie chaque kayak pour mieux les gérer 
+
+    ######################## On a pas du tout besoin de les gérer individuellement, donc je sais pas si c utile ##########################
+
     cur.execute("""
     CREATE TABLE IF NOT EXISTS kayak(
         id_kayak INTEGER PRIMARY KEY AUTOINCREMENT,
         type INT,
-        etat TEXT DEFAULT 'disponible'
+        etat TEXT DEFAULT 'disponible' ------------------- on fait pas de la gestion en live des nombres de kayaks dispo
         )
     """)
     
@@ -65,6 +70,8 @@ def creer_base(h_ouverture : int, min_ouverture : int, h_fermeture : int, min_fe
         cur.execute("INSERT INTO kayak(type) VALUES (1)")
     for _ in range(nb_2places):
         cur.execute("INSERT INTO kayak(type) VALUES (2)")
+
+
 
     # --- Création de la table location ---
     # Ajout du parcours pour les calculs de retour
@@ -106,18 +113,18 @@ def jour_suivant() -> tuple[int, int, int] :
     Fait passer un jour dans la base de données
     """
     cur.execute("SELECT * FROM calendrier")
-    a, m, j = cur.fetchone() #met les 3 valeurs du tuple retournées pas la méthode fetchone (de la forme (annee, mois, jour)) dans les variables a, m et j.
+    a, m, j = cur.fetchone() # met les 3 valeurs du tuple retournées pas la méthode fetchone (de la forme (annee, mois, jour)) dans les variables a, m et j.
 
     #Année bissextile ou non
     if a % 4 == 0 :
-        if m in [1,3,5,7,9,11] : #erreurs dans le premier fichier python sur les mois à 31
+        if m in [1,3,5,7,8,10,12] : #erreurs dans le premier fichier python sur les mois à 31 #### les mois a 31 jours sont [1,3,5,7,8,10,12], dit moi si je me trompe
             max_j = 31
         if m == 2 :
             max_j = 29
         else :
             max_j = 30
     else :
-        if m in [1,3,5,7,9,11] :
+        if m in [1,3,5,7,8,10,12] :
             max_j = 31
         if m == 2 :
             max_j = 28
@@ -144,19 +151,23 @@ def ajouter_client(nom: str, prenom: str) -> int:
     """
     Renvoie l'ID du client s'il existe déjà, sinon l'ajoute et Renvoie le nouvel ID.
     """
-    #On cherche si client il y a
+    # On cherche si client il y a
     cur.execute("SELECT id_client FROM client WHERE nom = ? AND prenom = ?", (nom, prenom))
     resultat = cur.fetchone()
-    #Si le client existe deja (On considère que 2 personnes n'ont pas le même nom et prenom...)
+    # Si le client existe deja (On considère que 2 personnes n'ont pas le même nom et prenom...)
     if resultat is not None:
-        return resultat[0] #On renvoie l'id deja présent
-    #Sinon on l'ajoute
+        return resultat[0] # On renvoie l'id deja présent
+    # Sinon on l'ajoute
     else:
         cur.execute("INSERT INTO client (nom, prenom) VALUES (?, ?)", (nom, prenom))
-        cur.execute("SELECT MAX(id_client) FROM client")
+        cur.execute("SELECT MAX(id_client) FROM client") 
+
+        # cur.execute("SELECT id_client FROM client WHERE nom = ? AND prenom = ?", (nom, prenom))
+        # au cas ou le max n'est pas le dernier ajouté en cas de suppression de client
+
         id_cree = cur.fetchone()[0]
         con.commit()
-        return id_cree #On renvoie nouvel id
+        return id_cree # On renvoie le nouvel id
         
 def ajoute_resa(j_depart: int, m_depart: int, a_depart: int, h_depart: int, min_depart: int, nb_1place: int, nb_2places: int, parcours: int, id_client: int) -> None:
     """
@@ -165,15 +176,15 @@ def ajoute_resa(j_depart: int, m_depart: int, a_depart: int, h_depart: int, min_
     #Vérification de la cohérence de la date
     cur.execute("""SELECT annee, mois, jour FROM calendrier""")
     a_actu, m_actu, j_actu = cur.fetchone()
-    if (a_depart, m_depart, j_depart) < (a_actu, m_actu, j_actu):
+    if (a_depart, m_depart, j_depart) < (a_actu, m_actu, j_actu): # <= ? on veut accepter les réservations pour le jour même ?
         print("Réservation impossible")
         return
 
     #Vérification de l'heure de départ en fonction du parcours
-    if parcours == 0 and h_depart > 15 :
+    if parcours == 0 and h_depart > 15 : # si 15h00 ?          ... and (h_depart > 15 or (h_depart == 15 and min_depart > 0))
         print("Parcours de 6km non autorisé après 15h")
         return
-    elif parcours == 1 and h_depart > 14:
+    elif parcours == 1 and h_depart > 14: # si 14h00 ?          ... and (h_depart > 14 or (h_depart == 14 and min_depart > 0))
         print("Parcours de 10km non autorisé après 14h")
         return
 
@@ -183,8 +194,24 @@ def ajoute_resa(j_depart: int, m_depart: int, a_depart: int, h_depart: int, min_
         print(f"Le client numéro : {id_client} n'existe pas.")
         return
 
-    #Vérification du stock de la JOURNEE (on peut prendre au max 50 réservations logiquement)
-    # A FAIRE !!!
+    #Vérification du stock de la JOURNEE (on peut prendre au max 100 réservé logiquement)
+    cur.execute("""
+        SELECT SUM(nb_1place), SUM(nb_2places) 
+        FROM location 
+        WHERE a_depart = ? AND m_depart = ? AND j_depart = ?""", 
+        (a_depart, m_depart, j_depart))
+    resultat = cur.fetchone()
+    # Verification de valeurs None
+    if resultat[0] is None:
+        resultat[0] = 0
+    if resultat[1] is None:
+        resultat[1] = 0
+    # Comparaison avec le stock total
+    if resultat[0] < nb_1place or resultat[1] < nb_2places:
+        print("Réservation impossible, pas assez de kayaks disponibles")
+        return
+
+    # Insertion de la réservation
     cur.execute("""
         INSERT INTO location (id_client, nb_1place, nb_2places, parcours, a_depart, m_depart, j_depart, h_depart, min_depart) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)""",
         (id_client, nb_1place, nb_2places, parcours, a_depart, m_depart, j_depart, h_depart, min_depart))
@@ -195,11 +222,24 @@ def supprime_resa(id_location : int, j_depart : int, m_depart : int, a_depart : 
     """
     Supprime une location si la date n'a pas été dépassée
     """
-    #Si location est plus ancienne que la date -> On ne peut pas la supprimer
+
+    # est-ce que la réservation existe
+    cur.execute(f"""SELECT * FROM location WHERE id_location = {id_location}""")
+    if cur.fetchone() is None :
+        print("Cette réservation n'existe pas")
+        return
+
+    # est-ce que on considère que l'on peut supprimer une réservation le jour même ?
+
+    # Si location est plus ancienne que la date -> On ne peut pas la supprimer
     cur.execute("""SELECT * FROM calendrier""")
     if cur.fetchone() <= (a_depart, m_depart, j_depart) : #Verifie si la date est plus petite qua la date de la location
         cur.execute(f"""DELETE FROM location WHERE id_location = {id_location}""")
         con.commit()
+        print("Réservation supprimée")
+    else :
+        print("Impossible de supprimer une réservation passée")
+
         
 def retour_kayaks2places(j_depart: int, m_depart: int, a_depart: int):
     """
@@ -217,7 +257,7 @@ def retour_kayaks2places(j_depart: int, m_depart: int, a_depart: int):
     rows = cur.fetchall()
     #L'employé passe toutes les 30 min
     #Fin du parcours 0 : 12h30, 13h30...
-    ramassage0 = [(12 + i, 30) for i in range(6)]
+    ramassage0 = [(12 + i, 30) for i in range(7)]
     #Fin du parcours 1 : 13h00, 14h00...
     ramassage1 = [(13 + i, 0) for i in range(6)]
 
@@ -237,7 +277,9 @@ def retour_kayaks2places(j_depart: int, m_depart: int, a_depart: int):
     dict_parcours0 = {k:0 for k in range(len(ramassage1))} #On met ramassage1 au lieu de ramassage0 plutôt
     i = 0
     #Ok mais si un kayak arrive après 17h30 a l'arrivée du parcours 0 ou après 18h a l'arrivée du parcours 1 on risque d'vaoir une erreur d'index il me semble (je suis pas sur mais a verifier)
-    while i < len(parcours0):
+    # j'ai ajouté une condition pour eviter l'erreur d'index
+    # on peut rajouter un passage de l'employé à 18h30 pour ramasser les kayaks du parcours 0 arrivant après 17h30
+    while i < len(parcours0) and j < len(ramassage0):
         if parcours0[i][2:4] <= ramassage0[j]:
             dict_parcours0[j] += parcours0[i][0]
             i += 1
@@ -249,16 +291,18 @@ def retour_kayaks2places(j_depart: int, m_depart: int, a_depart: int):
     j = 0
     dict_parcours1 = {k:0 for k in range(len(ramassage0))}
     i = 0
-    while i < len(parcours1):
+    while i < len(parcours1) and j < len(ramassage0):
         if parcours1[i][2:4] <= ramassage1[j]:
             dict_parcours1[j] += parcours1[i][0]
             i += 1
         else :
             j += 1
-            dict_parcours1[j] = 0
+            
     resultat1 = [ramassage1[k] + (dict_parcours1[k],) for k in range(len(ramassage1))]
 
     return (resultat0, resultat1)
+
+
 # --- A arranger ---       
 def retour_kayaks2places(j_depart : int, m_depart : int, a_depart : int) :
